Updates SM visualization TimeSeries to the composition API
